Fix float to double

+++ src/Transformable/Primitive/Sphere.cpp
@@ -8,7 +8,7 @@
 #include <cmath>
 #include "Sphere.hpp"
 
-Transformable::Primitive::Sphere::Sphere(Point3f pos, float _ray, Material::IMaterial material) : Transformable::Primitive::APrimitive(material, pos, {0, 0, 0}), _ray(_ray)
+Transformable::Primitive::Sphere::Sphere(Point3f pos, double _ray, Material::IMaterial material) : Transformable::Primitive::APrimitive(material, pos, {0, 0, 0}), _ray(_ray)
 {
 }
 
@@ -16,7 +16,7 @@ Transformable::Primitive::Sphere::Sphere(Point3f pos, float _ray, Material::IMat
 std::tuple<bool, Display::Color> Transformable::Primitive::Sphere::checkHit(std::unique_ptr<Raytracer::IVector> &vector)
 {
     Point3f vectorPos = vector->getPos();
-    float distance = sqrt(pow(vectorPos.x - _pos.x, 2) + pow(vectorPos.y - _pos.y, 2) + pow(vectorPos.z - _pos.z, 2));
+    double distance = sqrt(pow(vectorPos.x - _pos.x, 2) + pow(vectorPos.y - _pos.y, 2) + pow(vectorPos.z - _pos.z, 2));
 
     if (distance <= _ray) {
         return std::make_tuple(true, Display::Color{1, 1, 1});
